Migrations changed to make them more readable. More fields have been added to the tale user. Table and models created for medical specialities, countries and phone prefixes. More translation added, and the translations are noew stored as an user field (meaning, emails with translations for example). Fixed some errors regarding the intlTellInput plugin. Users phones are now added with prefixes. More valitations have been added, including the password one to make it more secure.

+++ database/migrations/2020_03_26_192511_create_roles_permissions.php
@@ -15,12 +15,10 @@ class CreateRolesPermissions extends Migration
     {
         Schema::create('roles_permissions', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->smallInteger('role_id')->unsigned()->comment('Associated role');
+            $table->bigInteger('role_id')->unsigned()->comment('Associated role');
             $table->bigInteger('permission_id')->unsigned()->comment('Associated permission');
             $table->integer('activated')->unsigned()->comment('Kind of permission. 0:Not activated/1:Activated');
-            // $table->string('value_name', 100)->comment('Permission name kind');
             $table->softDeletes();
-            // $table->timestamps();     
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP')); 
         });
